Move Target interface to SPI
