+++ pact/provider/core/src/main/java/org/arquillian/algeron/pact/provider/core/HttpTargetResourceProvider.java
@@ -1,6 +1,6 @@
 package org.arquillian.algeron.pact.provider.core;
 
-import org.arquillian.algeron.pact.provider.core.httptarget.Target;
+import org.arquillian.algeron.pact.provider.spi.Target;
 import org.jboss.arquillian.core.api.Instance;
 import org.jboss.arquillian.core.api.annotation.Inject;
 import org.jboss.arquillian.test.api.ArquillianResource;
